build: fix plugin for semantic release
