+++ tools/semantic-plugins/copy-package-info.mjs
@@ -1,0 +1,32 @@
+import * as fs from 'fs';
+import * as path from 'path';
+
+// const fs = require('fs');
+// const path = require('path');
+
+/* The plugin updates dependencies between packages based on the generated semantic-release version */
+
+export function prepare(pluginConfig, { logger }) {
+  const originalPackageJson = JSON.parse(
+    fs.readFileSync(pluginConfig.original, { encoding: 'utf-8' })
+  );
+
+  for (const packageDir of pluginConfig.packages) {
+    logger.log('Updating package.json of %s', packageDir);
+
+    const packageJson = JSON.parse(
+      fs.readFileSync(`${path.join(packageDir, 'package.json')}`, {
+        encoding: 'utf-8',
+      })
+    );
+
+    for (const key of pluginConfig.keys) {
+      packageJson[key] = originalPackageJson[key];
+    }
+
+    fs.writeFileSync(
+      `${path.join(packageDir, 'package.json')}`,
+      JSON.stringify(packageJson, undefined, 2)
+    );
+  }
+}
